feat: refactorizar la lógica de campañas, mover tipos y funciones a nuevas ubicaciones, y crear un hook para la gestión del formulario de creación de campañas con el objetivo de cumplir mejor con la arquitectura FSD
