+++ src/shared/api/campaign.ts
@@ -1,0 +1,5 @@
+export interface Campaign {
+  id: string
+  name: string
+  variant: 'existing' | 'new'
+}
