+++ src/views/campaigns/model/campaignStore.ts
